v4.0.0: decimal money types + opt-in tolerance for malformed rows
BREAKING — money values are decimal, not float/double.

Every rate/amount/fee property that was float? or double? is now decimal?:
TransferData.Rate, DeployData.Rate, DelegatorRewardData.Rate,
ValidatorRewardData.Rate, ValidatorData.Fee, RateData.Amount,
FTRateData.Amount, FTDexRateData.Amount, FTDailyRateData.Amount,
FTDailyDexRateData.Amount + .FtDailyRate, SwapData.Token0FtRate +
.Token1FtRate, and SupplyData's four issuance figures. 17 properties, 12 files.

Binary floating point cannot represent decimal fractions exactly, so these
arrived subtly wrong and the error compounded when a rate was multiplied by a
token amount. Single precision carries ~7 significant digits — enough to shift a
large transfer's fiat value by a fraction of a cent, systematically rather than
randomly, in whatever ledger the caller is keeping. This repo's own SupplyData
fixture already pinned issuance figures to 19 significant digits, which double
could never have round-tripped; those assertions just needed decimal literals.

Non-money floating point is deliberately untouched: latitude/longitude,
validator performance scores, and socket reconnect tuning stay double.

NEW — CasperCloudClientConfig.TolerateMalformedRows (default false).

When enabled, the envelope is deserialized with its rows detached and the rows
are then converted individually, so one unusable row costs only itself and the
loss is reported via the new SkippedItemCount on PaginatedResponse and
ListResponse. A malformed envelope still throws: a response whose envelope is
broken cannot be trusted at all, and tolerating rows must not quietly downgrade
that into a partial page.

Per-row conversion rather than a serializer Error handler — a handler scoped to
data[...] paths looked equivalent but is not: handling a member-level error tells
Newtonsoft to leave that member at its default and KEEP the row, yielding a
half-populated object and a skip count of zero. A test caught this.

Intended for pollers that must keep making progress. Treat any non-zero
SkippedItemCount as an incomplete page and do not advance a high-water mark past
it. Default off, so behaviour is unchanged for existing callers.

Tests: 5 new offline tests in CSPRCloudNetTolerantRowsTests.cs, using the
client's injectable HttpClient with a stub handler — no API key, no network.
Offline suite 45/45. Live integration suite: 38 failures vs 37 on unmodified
master; the single delta is GetBlocksAsync_ReturnsExpectedDataWithProposerAccountInfo,
which asserts the latest testnet block's proposer happens to have registered
account info — live-data chance, and unreachable from this diff since the tests
leave TolerateMalformedRows off, making the deserialize path byte-identical to
before. No JSON exceptions in either run.

Docs: README, llms.txt, CLAUDE.md (its "Current release" line was stale at
v2.9.0) and the csproj release notes.

+++ CSPR.Cloud.Net/Objects/Delegate/DelegatorRewardData.cs
@@ -45,7 +45,7 @@ namespace CSPR.Cloud.Net.Objects.Delegate
         /// Rate of the reward.
         /// </summary>
         [JsonProperty("rate")]
-        public float? Rate { get; set; }
+        public decimal? Rate { get; set; }
 
         /// <summary>
         /// Additional account information of the delegator.
